+++ modules/hanna-sprinkles/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. -->
+- feat: Add component `MainMenu2`
 - feat: Simplify `Layout` sprinkle to not handle mobile menu toggling
 - feat: Update `MainMenu` to use `MobileMenuToggler`
 - feat: Globally apply `setDefaultLanguage` based on page's `lang` attribute